pyproject.toml
[tool.poetry]
name = "kaggle-1"
version = "0.1.0"
description = ""
authors = ["Afanasyev Maxim, Azizov Said, Zhdanov Maxim"]
packages = [{include = "kekpipe", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.10"
kaggle = "^1.5.13"
pandas = "^2.0.0"
xztrainer = "^0.9.5"
scikit-learn = "^1.2.2"
light-the-torch = "^0.7.2"
transformers = "^4.27.4"
fastparquet = "^2023.2.0"
madgrad = "^1.3"
tensorboard = "^2.12.1"
onnx = "^1.13.1"
onnx2tf = "^1.9.2"
tensorflow-io-gcs-filesystem = "^0.32.0"
tensorflow = "^2.12.0"
onnxruntime = "^1.14.1"
onnxsim = "^0.4.19"
sne4onnx = "^1.0.11"
tflite-runtime = "2.9.1"
nobuco = "^0.1.1"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^3.6.3"
plotly = "^5.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
